refactor sgrep py (#221)
* refactor sgrep.py
* python workflow not relevant
* no more requirements installation
* cleanup imports, move config resolution to separate file (#222)
