+++ sgrep_lint/constants.py
@@ -1,0 +1,6 @@
+RCE_RULE_FLAG = "--dangerously-allow-arbitrary-code-execution-from-rules"
+RULES_KEY = "rules"
+ID_KEY = "id"
+
+DEFAULT_SGREP_CONFIG_NAME = "sgrep"
+DEFAULT_CONFIG_FILE = f".{DEFAULT_SGREP_CONFIG_NAME}.yml"
