+++ sgrep_lint/sgrep-lint
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+# sgrep-lint single executable must be run from the folder with all the shared objects
+
+THIS_DIR="$(dirname "$(realpath "$0")")";
+LD_PRELOAD="/bin/sgrep-lint-files:/usr/local/bin/sgrep-lint-files" ${THIS_DIR}/sgrep-lint-exe "$@"
