+++ sgrep_lint/tests/lint/stupid.js
@@ -1,0 +1,3 @@
+var x = 1;
+
+console.log(x == x)
