+++ sgrep_lint/tests/python/bad.yaml
@@ -4,7 +4,7 @@ rules:
       - pattern-inside:
         - pattern: $X == $X
         - pattern: $X != $X
-        - patterns:          
+        - patterns:
           - pattern-inside: |
               def __init__(...):
                   ...
