--- testlint/python/eqeq.expected.json
@@ -1,1 +1,0 @@
-{"results": [{"check_id": "testlint.python.eqeq-is-bad", "path": "tests/lint/stupid.py", "start": {"line": 3, "col": 12}, "end": {"line": 3, "col": 26}, "extra": {"message": "useless comparison operation `a+b == a+b` or `a+b != a+b`; possible bug?", "metavars": {"$X": {"start": {"line": 3, "col": 12, "offset": 55}, "end": {"line": 3, "col": 17, "offset": 60}, "abstract_content": "a+b"}}}}, {"check_id": "testlint.python.javascript-basic-eqeq-bad", "path": "tests/lint/stupid.js", "start": {"line": 3, "col": 13}, "end": {"line": 3, "col": 19}, "extra": {"message": "useless comparison", "metavars": {"$X": {"start": {"line": 3, "col": 13, "offset": 24}, "end": {"line": 3, "col": 14, "offset": 25}, "abstract_content": "x"}}}}]}
