--- tests/lint/stupid.js
@@ -1,3 +1,0 @@
-var x = 1;
-
-console.log(x == x)
